[Doc] Promote Metafunction from data type to concept, and document as such
